csa_backend/requirements.txt
transformers==4.35.2 
torch==2.2.1 
fastapi==0.115.0
python-multipart==0.0.9
uvicorn==0.32.0
#full text
requests==2.31.0
beautifulsoup4==4.12.3
sentence-transformers==2.7.0
numpy==1.26.4
selenium==4.17.2
markdown==3.6
openai==1.107.3
openai-agents==0.3.0
pinecone==6.0.2
html2text==2025.4.15
python-dotenv==1.1.0
pydantic-settings
apscheduler==3.11.0
webdriver-manager==4.0.2
supabase==2.15.0
backoff==2.2.1
tenacity==8.5.0
mailersend==0.6.0
email_validator==2.2.0
regex==2024.11.6
pytz==2025.2
pyyaml==6.0.2
RapidFuzz==3.13.0
vapi_server_sdk==1.5.1
stripe==8.10.0


fastapi-cache2[redis]==0.2.1
redis==4.6.0

